+++ kompressor-js/src/commonMain/kotlin/com/ensody/kompressor/js/JsInterop.kt
@@ -1,0 +1,13 @@
+package com.ensody.kompressor.js
+
+
+internal interface JsCompressionInterop {
+    val desiredSize: Int
+    suspend fun write(chunk: ByteArray, offset: Int, length: Int)
+    suspend fun close()
+    suspend fun read(): ByteArray?
+    suspend fun cancel()
+    suspend fun awaitReady()
+}
+
+internal expect fun createJsCompressionInterop(format: String, isCompression: Boolean): JsCompressionInterop