+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Added `kompressor-js` module providing support for JS and WasmJS targets using the browser's Compression Streams API.
+- Added `ExperimentalCompressionFormat` annotation to mark formats that might not be supported on all platforms.
+
 ## 0.4.5
 
 * Fixed Android bundling duplicate .so files from NativeBuilds.